third_party/WebKit/LayoutTests/inspector/debugger/async-callstack-middle-run-expected.txt

Tests that capturing asynchronous call stacks in debugger works if started after some time since the page loads.

Debugger was enabled.
Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) onClick (async-callstack-middle-run.html:55)
    1) clickedOnCapturing (async-callstack-middle-run.html:39)
    2) onClick (async-callstack-middle-run.html:52)
    3) clickedOffCapturing (async-callstack-middle-run.html:38)
    4) onClick (async-callstack-middle-run.html:49)
    5) clickedOnCapturing (async-callstack-middle-run.html:39)
    6) onClick (async-callstack-middle-run.html:46)
    7) clickedOffCapturing (async-callstack-middle-run.html:38)
    8) timeoutOnCapturing (async-callstack-middle-run.html:32)
    [setTimeout]
    0) timeoutOffCapturing (async-callstack-middle-run.html:21)

Call stack:
    0) onClick (async-callstack-middle-run.html:55)
    1) clickedOnCapturing (async-callstack-middle-run.html:39)
    2) onClick (async-callstack-middle-run.html:52)
    3) clickedOffCapturing (async-callstack-middle-run.html:38)
    4) onClick (async-callstack-middle-run.html:49)
    5) clickedOnCapturing (async-callstack-middle-run.html:39)
    6) timeoutOnCapturing (async-callstack-middle-run.html:33)
    [setTimeout]
    0) timeoutOffCapturing (async-callstack-middle-run.html:21)

Call stack:
    0) onImageError (async-callstack-middle-run.html:76)
    1) imageErrorOffCapturing (async-callstack-middle-run.html:60)

Call stack:
    0) onImageError (async-callstack-middle-run.html:79)
    1) imageErrorOnCapturing (async-callstack-middle-run.html:61)
    [Node.addEventListener("error")]
    0) timeoutOnCapturing (async-callstack-middle-run.html:30)
    [setTimeout]
    0) timeoutOffCapturing (async-callstack-middle-run.html:21)

Call stack:
    0) timeoutOnCapturing (async-callstack-middle-run.html:26)
    [setTimeout]
    0) timeoutOffCapturing (async-callstack-middle-run.html:21)

Debugger was disabled.
